task_1/scripts/requirements.txt
py-solc-x==2.0.3
web3==7.9.0
python-dotenv==1.0.1
